+++ src/coker/interfaces.py
@@ -17,3 +17,17 @@ class SymbolicCallable(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def lower(self):
         raise NotImplementedError
+
+
+class FunctionSignatureValue(metaclass=abc.ABCMeta):
+    """Value that declares the shapes of a callable function."""
+
+    @abc.abstractmethod
+    def input_shape(self):
+        """Return declared input shapes."""
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def output_shape(self):
+        """Return declared output shapes."""
+        raise NotImplementedError